Refactor play layer by compartmentalising into more focused modules
